refactor: rename useDebounce.ts to match the naming convention
